+++ pkg/channels/slack/interactions.go
@@ -308,6 +308,8 @@ func (a *Adapter) handleDecision(ctx context.Context, slackChannel, threadID, me
 	turnCtx, done := a.registerTurn(ctx, threadID)
 	defer done()
 
+	a.logTurnDispatch(msg, slackUser, true)
+
 	deltas, err := a.gw.SendCompletion(turnCtx, ref, msg)
 	if err != nil {
 		a.storePendingTask(threadID, task)